Step1 - 즐겨찾기 기능 완성 (#679)
* feat: 이전 프로젝트 초기화

* docs: 요구사항 정리

* feat: 비로그인 즐겨찾기 기능 사용시 401 응답

* chore: 즐겨찾기 기능 개발 중 Path 리팩터링을 위해 잠시 멈춤

* refactor: Paths 도메인에서 책임져야 할 부분들을 서비스에서 도메인으로 이동

* test: 경로 부족한 테스트 추가

* feat: 즐겨찾기 추가, 즐겨찾기 조회 기능 추가

* feat: 즐겨찾기 삭제 기능 추가

+++ src/test/java/nextstep/utils/DatabaseCleanup.java
@@ -1,5 +1,7 @@
 package nextstep.utils;
 
+import nextstep.line.acceptance.LineAcceptanceTest;
+import nextstep.path.acceptance.PathAcceptanceTest;
 import org.springframework.beans.factory.InitializingBean;
 import org.springframework.context.annotation.Profile;
 import org.springframework.stereotype.Service;
@@ -11,7 +13,7 @@ import javax.persistence.PersistenceContext;
 import java.util.List;
 import java.util.stream.Collectors;
 
-@Profile("test")
+@Profile("databaseCleanup")
 @Service
 public class DatabaseCleanup implements InitializingBean {
     @PersistenceContext
@@ -28,13 +30,28 @@ public class DatabaseCleanup implements InitializingBean {
     }
 
     @Transactional
-    public void execute() {
+    public void execute(Object testClass) {
         entityManager.flush();
         entityManager.createNativeQuery("SET REFERENTIAL_INTEGRITY FALSE").executeUpdate();
         for (String tableName : tableNames) {
-            entityManager.createNativeQuery("TRUNCATE TABLE " + tableName).executeUpdate();
-            entityManager.createNativeQuery("ALTER TABLE " + tableName + " ALTER COLUMN ID RESTART WITH 1").executeUpdate();
+            truncateTables(testClass, tableName);
         }
         entityManager.createNativeQuery("SET REFERENTIAL_INTEGRITY TRUE").executeUpdate();
+    }
+
+    private void truncateTables(Object testClass, String tableName) {
+        if (notTruncateAble(testClass, tableName)) return;
+        entityManager.createNativeQuery("TRUNCATE TABLE " + tableName).executeUpdate();
+        entityManager.createNativeQuery("ALTER TABLE " + tableName + " ALTER COLUMN ID RESTART WITH 1").executeUpdate();
+    }
+
+    private static boolean notTruncateAble(Object testClass, String tableName) {
+        if (testClass.equals(LineAcceptanceTest.class) && tableName.equals("Station")) {
+            return true;
+        }
+        if (testClass.equals(PathAcceptanceTest.class) && tableName.equals("Station")) {
+            return true;
+        }
+        return false;
     }
 }
